+++ crates/sui-indexer-alt/src/handlers/mod.rs
@@ -13,7 +13,6 @@ pub(crate) mod kv_objects;
 pub(crate) mod kv_protocol_configs;
 pub(crate) mod kv_transactions;
 pub(crate) mod obj_info;
-pub(crate) mod obj_info_pruner;
 pub(crate) mod obj_versions;
 pub(crate) mod sum_displays;
 pub(crate) mod sum_packages;
