+++ crates/sui-indexer-alt/src/lib.rs
@@ -6,7 +6,6 @@ use bootstrap::bootstrap;
 use config::{ConsistencyConfig, IndexerConfig, PipelineLayer};
 use handlers::coin_balance_buckets::CoinBalanceBuckets;
 use handlers::coin_balance_buckets_pruner::CoinBalanceBucketsPruner;
-use handlers::obj_info_pruner::ObjInfoPruner;
 use handlers::{
     ev_emit_mod::EvEmitMod, ev_struct_inst::EvStructInst, kv_checkpoints::KvCheckpoints,
     kv_epoch_ends::KvEpochEnds, kv_epoch_starts::KvEpochStarts, kv_feature_flags::KvFeatureFlags,
@@ -59,8 +58,6 @@ pub async fn start_indexer(
     let PipelineLayer {
         sum_displays,
         sum_packages,
-        obj_info,
-        obj_info_pruner,
         coin_balance_buckets,
         coin_balance_buckets_pruner,
         cp_sequence_numbers,
@@ -73,6 +70,7 @@ pub async fn start_indexer(
         kv_objects,
         kv_protocol_configs,
         kv_transactions,
+        obj_info,
         obj_versions,
         tx_affected_addresses,
         tx_affected_objects,
@@ -193,11 +191,6 @@ pub async fn start_indexer(
     add_sequential!(SumPackages, sum_packages);
 
     add_consistent!(
-        ObjInfo::default(), obj_info;
-        ObjInfoPruner, obj_info_pruner
-    );
-
-    add_consistent!(
         CoinBalanceBuckets, coin_balance_buckets;
         CoinBalanceBucketsPruner, coin_balance_buckets_pruner
     );
@@ -211,6 +204,7 @@ pub async fn start_indexer(
     add_concurrent!(KvEpochStarts, kv_epoch_starts);
     add_concurrent!(KvObjects, kv_objects);
     add_concurrent!(KvTransactions, kv_transactions);
+    add_concurrent!(ObjInfo::default(), obj_info);
     add_concurrent!(ObjVersions, obj_versions);
     add_concurrent!(TxAffectedAddresses, tx_affected_addresses);
     add_concurrent!(TxAffectedObjects, tx_affected_objects);
